[X][Refactor] Rename schedule-widget to Hour-schedule-widget
