+++ src/ui/hour-schedule-widget/hour-schedule-item/hour-schedule-item.styles.ts
