+++ src/ui/hour-schedule-widget/hour-schedule-widget.container.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
-import { IForecastday, IWeather } from '../../common/types/IWeather';
-import ScheduleWidget from './schedule-widget';
+import { IForecastday } from '../../common/types/IWeather';
+import ScheduleWidget from './hour-schedule-widget';
 import { useSelector } from 'react-redux';
 import { IRootState } from '../../common/types/store';
 
@@ -8,7 +8,7 @@ interface Props {
   days: Array<IForecastday>;
 }
 
-const ScheduleWidgetContainer: FC<Props> = ({ days }) => {
+const HourScheduleWidgetContainer: FC<Props> = ({ days }) => {
   const dayIndex = useSelector(
     (state: IRootState) => state.weather.dayOfTheWeek
   );
@@ -16,4 +16,4 @@ const ScheduleWidgetContainer: FC<Props> = ({ days }) => {
   return <>{<ScheduleWidget schedule={days[dayIndex].hour} />}</>;
 };
 
-export default ScheduleWidgetContainer;
+export default HourScheduleWidgetContainer;
