+++ src/ui/hour-schedule-widget/hour-schedule-widget.helpers.ts
