+++ src/ui/hour-schedule-widget/hour-schedule-widget.styles.ts
